+++ src/dispatch/controllers.js
@@ -1,0 +1,67 @@
+'use strict';
+
+function DispatcherCtrl($state, Dispatcher) {
+
+  console.debug('called ctrl');
+  /*
+
+   resolvedPath will return a dictionary such as:
+
+   {
+   error: 'Some Error Condition'
+   schema: 'Some Schema Identifier'
+   data: {
+   viewName: the name of the view,
+   context: the context object,
+   parents: the parents array,
+   view: the dict returned by any custom view
+   items: sequence of children if it is a folder
+   ordering: if ordered folder, the ordering of the item ids
+   }
+
+   }
+
+   */
+
+  // First hande the case where resolvedPath says it couldn't
+  // find anything.
+
+  if (resolvedPath.error) {
+    // This should be a not found
+    $state.go('notfound');
+  }
+
+  var data = resolvedPath.data.data;
+  Dispatcher.context = data.context;
+  Dispatcher.viewName = data.viewName;
+  Dispatcher.parents = data.parents;
+
+  // Get the next state. Look in all the registered states at
+  // view_config information.
+  var matchingView = Dispatcher.resolveState(
+    Dispatcher.context, Dispatcher.viewName, Dispatcher.parents);
+  var nextState = matchingView.stateName;
+
+  if (nextState) {
+    $state.go(nextState);
+  } else {
+    // Dispatcher failed to find a matching view
+    $state.go('notfound');
+  }
+
+}
+
+function NotFoundCtrl($location) {
+  this.path = $location.path();
+}
+
+function ErrorCtrl($stateParams) {
+  this.toState = $stateParams.toState;
+  this.error = $stateParams.error;
+}
+
+
+angular.module('md.dispatch')
+  .controller('NotFoundCtrl', NotFoundCtrl)
+  .controller('ErrorCtrl', ErrorCtrl)
+  .controller('DispatcherCtrl', DispatcherCtrl);